requirements.txt
aniso8601==7.0.0
asgiref==3.2.4
aspy.yaml==1.3.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.5.3
certifi==2019.11.28
cfgv==2.0.1
chardet==3.0.4
coverage==5.0.4
coveralls==2.0.0
dj-database-url==0.5.0
Django==3.0.7
django-environ==0.4.5
django-filter==2.3.0
django-graphql-auth==0.3.10
django-graphql-jwt==0.3.0
django-heroku==0.3.1
docopt==0.6.2
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
gunicorn==20.0.4
identify==1.4.9
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
more-itertools==8.2.0
nodeenv==1.3.3
packaging==20.3
pep8==1.7.1
pluggy==0.13.1
pre-commit==1.21.0
promise==2.3
psycopg2==2.8.5
py==1.8.1
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.7
pytest==5.4.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
Rx==1.6.1
singledispatch==3.4.0.3
six==1.14.0
sqlparse==0.3.1
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
virtualenv==16.7.9
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.11.2
zipp==3.1.0
